DS2022_30642_Turdean_Bogdan_1_Backend/requirements.txt
alembic==1.7.7
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
banal==1.0.6
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
channels==4.0.0
charset-normalizer==2.0.12
cryptography==36.0.2
cycler==0.11.0
dataset==1.5.2
distlib==0.3.4
Django==4.1.3
django-cors-headers==3.13.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
docker==6.0.1
docker-build==0.1.2
filelock==3.6.0
fonttools==4.31.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
kiwisolver==1.4.2
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
mysqlclient==2.1.1
numpy==1.22.3
outcome==1.1.0
packaging==21.3
pandas==1.4.1
Pillow==9.1.0
platformdirs==2.5.1
psycopg2==2.9.5
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2020.1
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
sqlparse==0.3.1
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
websocket-client==1.4.2
wsproto==1.1.0
